add GetBlock api
