+++ src/AetherLink.Worker/appsettings.json
@@ -30,6 +30,14 @@
             "NoApiKeyPerSecondRequestLimit": 1,
             "Weight": 60,
             "ApiKey": ""
+          },
+          "GetBlock": {
+            "Url":"https://go.getblock.io/",
+            "FreeApi": true,
+            "ApiKeyPerSecondRequestLimit": 6,
+            "NoApiKeyPerSecondRequestLimit": 1,
+            "Weight": 40,
+            "ApiKey": "175ded7b88dd486088bfe0712e321a29"
           }
         }
       }
